requirements.txt
aiogram==2.24
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
exceptiongroup==1.1.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
itsdangerous==2.1.2
magic-filter==1.0.9
MarkupSafe==2.1.1
moviepy==1.0.3
multidict==6.0.4
mutagen==1.46.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pafy==0.5.5
Pillow==9.4.0
pip-review==1.3.0
proglog==0.1.10
pyaes==1.6.1
pycryptodomex==3.17
pydantic==1.10.2
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytube==12.1.2
pytz==2022.7.1
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.15
webdriver-manager==3.8.5
websockets==10.4
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
youtube-dl==2021.12.17
zipp==3.15.0
